% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_geneset.r
\name{display_genesets}
\alias{display_genesets}
\title{This function displays the characteristics of the Hazard Ratio gene sets we have curated/included.}
\usage{
display_genesets()
}
\value{
Hazard Ratio gene sets we have curated/included.
}
\description{
This function displays the characteristics of the Hazard Ratio gene sets we have curated/included.
}
\examples{
library(HazardGeneEnrichment)
display_genesets()


}
