% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaplot_hazard.r
\name{gseaplot_hazard}
\alias{gseaplot_hazard}
\title{This function is used to display the results of gene sets enriched for hazard ratio.}
\arguments{
\item{x}{is result of hazardenrichment.r}

\item{geneSetID}{Gene sets to display}

\item{pvalue_table}{Whether to add pvalue_table}
}
\value{
Plot of gene set enrichment on Hazard Ratio-ranked gene sets.
}
\description{
This function is used to display the results of gene sets enriched for hazard ratio.
}
\examples{
library(HazardGeneEnrichment)
#my_gsea_results is result of gseaplot_hazard.r
gseaplot_hazard(my_gsea_results, geneSetID = "My_GeneSet_1", title = "GSEA Plot for My_GeneSet_1")
gseaplot2(my_gsea_results, geneSetID = c("My_GeneSet_1", "My_GeneSet_2"),
title = "GSEA Plot for Custom Gene Sets")
}
