% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazardenrichment.r
\name{HazardEnrichment}
\alias{HazardEnrichment}
\title{This function fits a Cox proportional hazards model and extracts
the hazard ratios for specified genes.}
\usage{
HazardEnrichment(geneList, TERM2GENE)
}
\arguments{
\item{geneList}{The reference gene sets we need to enrich to,for example 'TCGA-BRCA'}

\item{TERM2GENE}{The gene sets to be analyzed.}
}
\value{
my_gsea_results
}
\description{
This function fits a Cox proportional hazards model and extracts
the hazard ratios for specified genes.
}
\examples{
library(HazardGeneEnrichment)
HazardEnrichment('TCGA-BRCA',TERM2GENE)
}
